1.0.360.2565

+++ M2k/M2kCommand.cs
@@ -504,7 +504,7 @@ namespace M2kClient
             //Adjusting out any reclaim from the system and then adjusting it back in as the raw compound
             if (wipRecord.IsReclaim == SFW.Model.Enumerations.Complete.Y)
             {
-                var test = Convert.ToInt32(Math.Round(Convert.ToDouble(wipRecord.ReclaimQty) * wipRecord.ReclaimAssyQty, 0, MidpointRounding.AwayFromZero));
+                var test = Convert.ToInt32(Math.Round(Convert.ToDecimal(wipRecord.ReclaimQty) * wipRecord.ReclaimAssyQty, 0, MidpointRounding.AwayFromZero));
                 //Adjustment out
                 InventoryAdjustment(wipRecord.Submitter,
                     $"{wipRecord.ReclaimReference}*{wipRecord.WipWorkOrder.OrderNumber}",
@@ -520,7 +520,7 @@ namespace M2kClient
                     wipRecord.ReclaimParent,
                     AdjustCode.REC,
                     'A',
-                    Convert.ToInt32(Math.Round(Convert.ToDouble(wipRecord.ReclaimQty) * wipRecord.ReclaimAssyQty, 0, MidpointRounding.AwayFromZero)),
+                    Convert.ToInt32(Math.Round(Convert.ToDecimal(wipRecord.ReclaimQty) * wipRecord.ReclaimAssyQty, 0, MidpointRounding.AwayFromZero)),
                     "EXT-1",
                     connection);
             }
